minor updates and doxygen comments added

+++ src/tamm/types.h
@@ -16,10 +16,10 @@ namespace tamm {
 struct IrrepSpace;
 using Irrep = StrongNum<IrrepSpace, uint32_t>;
 struct SpinSpace;
-using Spin       = StrongNum<SpinSpace, uint32_t>;
+using Spin = StrongNum<SpinSpace, uint32_t>;
 struct SpatialSpace;
-using Spatial       = StrongNum<SpatialSpace, uint32_t>;
-using TensorRank = uint32_t;
+using Spatial    = StrongNum<SpatialSpace, uint32_t>;
+using TensorRank = size_t;
 struct OffsetSpace;
 using Offset = StrongNum<OffsetSpace, uint64_t>;
 struct BlockIndexSpace;
